Bootstrap implemented

+++ app/controllers/sections_controller.rb
@@ -1,7 +1,7 @@
 class SectionsController < ApplicationController
   protect_from_forgery :except => :all 
   
-  before_action :set_classroom, only: [:create,:destroy]
+  before_action :set_classroom, only: [:create,:destroy,:show]
   
   def index
     if params[:classroom_id]
@@ -12,28 +12,36 @@ class SectionsController < ApplicationController
       render json: @sections
     end
   end
-  
+ 
   #this will add a new section to the classroom upon which it is called
   def create
     last_section = last_section_find.section
-    new_section = (last_section[0].ord + 1).chr
-    @section = @classroom.sections.new(section: new_section)
-    if @section.save
-      render json: @section
+    if last_section[0].ord <= 75
+      new_section = (last_section[0].ord + 1).chr
+      @section = @classroom.sections.new(section: new_section)
+      if @section.save
+        render json: @section
+      else
+        render json: '{"error":"unable to create section"}'
+      end
     else
-      render json: '{"error":"unable to create section"}'
+        render json: '{"error":"Can only create uptp 10 sections"}'      
     end
+    
   end
   
   #this will remove the last section added in the particular classroom
   def destroy
     last_section = last_section_find
-    if last_section.destroy
-      render json: last_section
-    else
-      render json: '{"success":"no"}'
+    if last_section.section[0].ord > 65
+      if last_section.destroy
+        render json: last_section
+      else
+        render json: '{"success":"no"}'
+      end
+     else
+      render json: '{"success":"no","why":"can\' delete last section"}'   
     end
-    
   end
